pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["dlee <derekclee323@g.ucla.edu>"]

[tool.poetry.dependencies]
python = ">=3.8, <3.10"
requests = "^2.25.1"
selenium = "^3.141.0"
pytesseract = "^0.3.7"
numpy = "^1.20.1"
Pillow = "^8.1.2"
reportlab = "^3.5.66"
autocorrect = "^2.3.0"
scipy = "^1.6.2"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
